html/webappapis/dynamic-markup-insertion/document-write/module-delayed.html

<!doctype html>
<meta charset=utf-8>
<title>async document.write in a module</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(test => {
  // Expose {test} in the iframe for using the step_timeout helper.
  document.test = test;
  const iframe = document.createElement("iframe");
  iframe.src = "module-delayed-iframe.html";
  document.body.appendChild(iframe);
  iframe.onerror = test.unreached_func("Error loading iframe");
  let onLoadWasCalled = false;
  iframe.onload =  test.step_func(() => {
    assert_equals(iframe.contentDocument.body.textContent, "FAIL\n");
    // Don't call the event handler another time after document.write.
    iframe.onload = null;
    onLoadWasCalled = true;
  });
  document.addEventListener("testEnd", test.step_func_done(() => {
    assert_true(onLoadWasCalled);
    assert_equals(iframe.onload, null);
    assert_equals(iframe.contentDocument.body.textContent, "PASS\n");
  }));
});
</script>
